requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

numpy>=1.24.0
pandas>=2.0.0

scikit-learn>=1.3.0
segmentation-models-pytorch>=0.3.3

opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

matplotlib>=3.7.0
seaborn>=0.12.0

tqdm>=4.65.0
pyyaml>=6.0
kaggle>=1.5.16

jupyter>=1.0.0
ipykernel>=6.25.0
